phase1_input_security/adversarial_examples/archive/requirements_llava.txt
# Dependencias adicionales para testing con LLaVA
torch>=2.0.0
transformers>=4.37.0
accelerate>=0.20.0
safetensors>=0.3.0

# Para imágenes y procesamiento
Pillow>=10.0.0
